+++ tests/test_node_monitor.py
@@ -28,8 +28,7 @@ mock_node_provider_db.get_subscribers_as_dict.return_value = \
       'notify_on_status_change': True,
       'notify_email': True,
       'notify_slack': True,
-      'notify_telegram_chat': True,
-      'notify_telegram_channel': True # Deprecated: column not used
+      'notify_telegram': True,
       }
     }
 mock_node_provider_db.get_node_labels_as_dict.return_value = \
@@ -38,15 +37,12 @@ mock_node_provider_db.get_node_labels_as_dict.return_value = \
 mock_node_provider_db.get_emails_as_dict.return_value = \
     {'rbn2y-6vfsb-gv35j-4cyvy-pzbdu-e5aum-jzjg6-5b4n5-vuguf-ycubq-zae':
      ['test_recipient@gmail.com']}
-mock_node_provider_db.get_channels_as_dict.return_value = \
+mock_node_provider_db.get_slack_channels_as_dict.return_value = \
     {'rbn2y-6vfsb-gv35j-4cyvy-pzbdu-e5aum-jzjg6-5b4n5-vuguf-ycubq-zae':
-     {'id': 1, 
-      'node_provider_id': 'rbn2y-6vfsb-gv35j-4cyvy-pzbdu-e5aum-jzjg6-5b4n5-vuguf-ycubq-zae', 
-      'slack_channel_name': '#node-monitor', 
-      'telegram_chat_id' : '5734534558', 
-      'telegram_channel_id': '-1001925583150' # Deprecated: column not used
-      } 
-    }
+        ['#node-monitor']}
+mock_node_provider_db.get_telegram_chats_as_dict.return_value = \
+    {'rbn2y-6vfsb-gv35j-4cyvy-pzbdu-e5aum-jzjg6-5b4n5-vuguf-ycubq-zae':
+        ['5734534558']}
 mock_node_provider_db.get_node_providers_as_dict.return_value = \
     {'64xe5-tx2s3-4gjmj-pnozr-fejw2-77y5y-rhcjk-glnmx-62brf-qin5q-pqe': '', 
      '7k7b7-4pzhf-aivy6-y654t-uqyup-2auiz-ew2cm-4qkl4-nsl4v-bul5k-5qe': '1G', 
@@ -105,8 +101,8 @@ def test_control():
     # test broadcast_alerts()
     nm.broadcast_alerts()
     assert mock_email_bot.send_emails.call_count == 0
-    assert mock_slack_bot.send_message.call_count == 0
-    assert mock_telegram_bot.send_message.call_count == 0
+    assert mock_slack_bot.send_messages.call_count == 0
+    assert mock_telegram_bot.send_messages.call_count == 0
     mock_node_provider_db.reset_mock()
     mock_email_bot.reset_mock()
     mock_slack_bot.reset_mock()
@@ -115,8 +111,8 @@ def test_control():
     # test broadcast_status_report()
     nm.broadcast_status_report()
     assert mock_email_bot.send_emails.call_count == 1
-    assert mock_slack_bot.send_message.call_count == 1
-    assert mock_telegram_bot.send_message.call_count == 1
+    assert mock_slack_bot.send_messages.call_count == 1
+    assert mock_telegram_bot.send_messages.call_count == 1
     mock_node_provider_db.reset_mock()
     mock_email_bot.reset_mock()
     mock_slack_bot.reset_mock()
@@ -148,8 +144,8 @@ def test_one_node_bounce():
     # test broadcast_alerts()
     nm.broadcast_alerts()
     assert mock_email_bot.send_emails.call_count == 0
-    assert mock_slack_bot.send_message.call_count == 0
-    assert mock_telegram_bot.send_message.call_count == 0
+    assert mock_slack_bot.send_messages.call_count == 0
+    assert mock_telegram_bot.send_messages.call_count == 0
     mock_node_provider_db.reset_mock()
     mock_email_bot.reset_mock()
     mock_slack_bot.reset_mock()
@@ -178,8 +174,8 @@ def test_two_nodes_down():
     # test broadcast_alerts()
     nm.broadcast_alerts()
     assert mock_email_bot.send_emails.call_count == 1
-    assert mock_slack_bot.send_message.call_count == 1
-    assert mock_telegram_bot.send_message.call_count == 1
+    assert mock_slack_bot.send_messages.call_count == 1
+    assert mock_telegram_bot.send_messages.call_count == 1
     mock_node_provider_db.reset_mock()
     mock_email_bot.reset_mock()
     mock_slack_bot.reset_mock()